Refactoring EnvelopeNetworkRx network layer extensions.
Summary:
In order for the Rx network operation to be able to re-try a failed *request*, and not only keep returning whatever it received previously, the underlying implementation should actually call `network.request()` every time Rx operation is being subscribed to.
To achieve this behaviour, previous implementation of Rx extensions needed to be changed, so that the transition to Rx world happens *before* making a request.
Also, needed to take care of intermediate calls to `progress()` and `validate()` to be properly re-created every time the operation is created, i.e.:
```
network
    .rx.request()
    .validate()
    .progress()
    .response()
```